+++ test/Test/Codec/CBOR/Cuddle/CDDL/GeneratorSpec.hs
@@ -7,7 +7,7 @@ module Test.Codec.CBOR.Cuddle.CDDL.GeneratorSpec (spec) where
 import Codec.CBOR.Cuddle.CBOR.Gen (GenPhase, generateFromName)
 import Codec.CBOR.Cuddle.CBOR.Validator (validateCBOR)
 import Codec.CBOR.Cuddle.CDDL (Name)
-import Codec.CBOR.Cuddle.CDDL.CBORGenerator (GenEnv (..), runCBORGen)
+import Codec.CBOR.Cuddle.CDDL.CBORGenerator (GenConfig (..), runCBORGen)
 import Codec.CBOR.Cuddle.CDDL.CTree (CTreeRoot (..))
 import Codec.CBOR.Cuddle.CDDL.Resolve (MonoReferenced, MonoSimple, fullResolveCDDL)
 import Codec.CBOR.Cuddle.Huddle (Huddle, toCDDL)
@@ -30,6 +30,7 @@ import Test.Codec.CBOR.Cuddle.CDDL.Examples.Huddle (
   refTermExample,
   sizeBytesExample,
   sizeTextExample,
+  tagRangeExample,
   taggedUintExample,
  )
 import Test.Codec.CBOR.Cuddle.CDDL.Validator (expectInvalid, genAndValidateRule)
@@ -40,12 +41,12 @@ import Test.QuickCheck (Gen, Property, Testable (..), classify, counterexample)
 import Text.Pretty.Simple (pShow)
 
 generateCDDL :: CTreeRoot GenPhase -> Gen Term
-generateCDDL cddl = runAntiGen . runCBORGen env $ generateFromName "root"
+generateCDDL cddl = runAntiGen . runCBORGen cfg $ generateFromName "root"
   where
-    env =
-      GenEnv
-        { geRoot = cddl
-        , geTwiddle = True
+    cfg =
+      GenConfig
+        { gcRoot = cddl
+        , gcTwiddle = True
         }
 
 tryResolveHuddle :: HasCallStack => Huddle -> SpecM () (CTreeRoot MonoReferenced)
@@ -57,12 +58,12 @@ tryResolveHuddle huddle = do
 expectZapInvalidates :: CTreeRoot MonoReferenced -> Name -> Property
 expectZapInvalidates cddl name = property $ do
   let
-    env =
-      GenEnv
-        { geRoot = mapIndex cddl
-        , geTwiddle = True
+    cfg =
+      GenConfig
+        { gcRoot = mapIndex cddl
+        , gcTwiddle = True
         }
-  res@ZapResult {zrValue} <- zapAntiGenResult 1 . runCBORGen env $ generateFromName name
+  res@ZapResult {zrValue} <- zapAntiGenResult 1 . runCBORGen cfg $ generateFromName name
   let
     bs = toStrictByteString $ encodeTerm zrValue
     validationRes = validateCBOR bs name $ mapIndex cddl
@@ -99,6 +100,7 @@ spec = do
       genAndValidateHuddle "sizeBytes" sizeBytesExample
       genAndValidateHuddle "rangeList" rangeListExample
       genAndValidateHuddle "rangeMap" rangeMapExample
+      genAndValidateHuddle "tagRange" tagRangeExample
       xdescribe "Generator cannot reliably produce unique keys for maps" $ do
         genAndValidateHuddle "optionalMapExample" optionalMapExample
 
@@ -139,17 +141,31 @@ spec = do
             TBytes "\x01\x02\x03\xff" -> True
             TBytesI "\x01\x02\x03\xff" -> True
             _ -> False
+      tagRangeExampleCddl <- tryResolveHuddle tagRangeExample
+      prop "tagRange generator covers edges and middle" $ do
+        res <- generateCDDL $ mapIndex tagRangeExampleCddl
+        let tags = case res of
+              TList xs -> [t | TTagged t _ <- xs]
+              TListI xs -> [t | TTagged t _ <- xs]
+              _ -> []
+            classifyTag t p =
+              classify (t == 1280 || t == 1400) "edge tag (1280 or 1400)" $
+                classify (t > 1280 && t < 1400) "middle tag (1281..1399)" p
+            base =
+              counterexample ("tags: " <> show tags) $
+                all (\t -> t >= 1280 && t <= 1400) tags
+        pure $ foldr classifyTag (property base) tags
 
   describe "Tagged bytes zapping" $ do
     taggedBytesCddl <- tryResolveHuddle taggedUintExample
-    let env =
-          GenEnv
-            { geRoot = mapIndex taggedBytesCddl
-            , geTwiddle = True
+    let cfg =
+          GenConfig
+            { gcRoot = mapIndex taggedBytesCddl
+            , gcTwiddle = True
             }
     prop "labels zapped result" $ do
       ZapResult {zrValue} <-
-        zapAntiGenResult 1 . runCBORGen env $ generateFromName "root"
+        zapAntiGenResult 1 . runCBORGen cfg $ generateFromName "root"
       let
         isBytes TBytes {} = True
         isBytes TBytesI {} = True